move bazaar fetchers as well
